Moved late interaction code to one file
